+++ src/api/v1/platform/platform.constant.ts
@@ -1,0 +1,6 @@
+export const success = {
+  PLATFORM_CREATED_SUCCESSFULLY: "Platform created successfully",
+  PLATFORM_FETCHED_SUCCESSFULLY: "Platform fetched successfully",
+  PLATFORM_UPDATED_SUCCESSFULLY: "Platform updated successfully",
+  PLATFORM_DELETED_SUCCESSFULLY: "Platform deleted successfully",
+};