+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to the User Service will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - Unreleased
+
+### Added
+
+- Add Platform Model & API Endpoints ([#145](https://github.com/ansopedia/user-service/issues/145))
+- Protect Permission Routes ([#69](https://github.com/ansopedia/user-service/issues/69))
+
 ## [1.3.1] - 2025-10-06
 
 ### Added